Basic mouse functionality.
Thanks to degs' research at http://www.68k.org/~degs/nextkeyboard.html
who documented the signal to the keyboard which was comprised of a
keyboard query (already used) and a mouse query which has been added.

This is currently very jumpy and occasionally starts going in the wrong
direction, but is a start at least.